pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "==3.12.8"
dependencies = [
    "fiona",
    "flask",
    "flask-cors",
    "flask-session",
    "mysql-connector-python",
    "numpy",
    "openpyxl",
    "pandas",
    "pyproj",
    "scipy",
    "rpy2",
    "shapely",
    "sqlalchemy",
    "tqdm",
    "sqlacodegen==3.0.0rc5",
    "passlib",
    "argon2-cffi",
    "watchdog",
    "pytest",
    "fpdf2",
    "matplotlib",
    "flask-executor",
    "cartopy",
    "cachelib",
    "freezegun>=1.5.5",
    "python-stripzip>=2.0.0",
]

[dependency-groups]
dev = []

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
